requirements.txt
requests
python-dotenv
sqlalchemy
pandas
matplotlib
